fix: clang warning/errors
now it is possible to compile using LLVM(clang) in visual studio

+++ src/framework/core/eventdispatcher.cpp
@@ -76,7 +76,7 @@ void EventDispatcher::poll()
             break;
         }
 
-        for (int_fast32_t i = -1; ++i < m_pollEventsSize;) {
+        for (int_fast32_t i = -1; ++i < static_cast<int_fast32_t>(m_pollEventsSize);) {
             const auto event = m_eventList.front();
             m_eventList.pop_front();
             if (isMainDispatcher) ul.unlock();
